backend/requirements.txt
python-socketio==5.2.0
